+++ routes/web.php
@@ -11,6 +11,7 @@
 |
 */
 
+
 // Authentication Routes...
 Route::get('login', [\App\Http\Controllers\Auth\LoginController::class,'showLoginForm'])->name('login');
 Route::post('login', [\App\Http\Controllers\Auth\LoginController::class,'login']);
